ticketvote: Add startStandard.
