+++ politeiad/backendv2/tstorebe/plugins/ticketvote/ticketvote.go
@@ -144,16 +144,14 @@ func (p *plugin) Write(tstore plugins.TstoreClient, token []byte, cmd, payload s
 	switch cmd {
 	case ticketvote.CmdAuthorize:
 		return p.cmdAuthorize(tstore, token, payload)
-		/*
-			case ticketvote.CmdStart:
-				return p.cmdStart(tstore, token, payload)
-			case ticketvote.CmdCastBallot:
-				return p.cmdCastBallot(tstore, token, payload)
+	case ticketvote.CmdStart:
+		return p.cmdStart(tstore, token, payload)
+		// case ticketvote.CmdCastBallot:
+		// return p.cmdCastBallot(tstore, token, payload)
 
-				// Internal plugin commands
-			case cmdStartRunoffSubmission:
-				return p.cmdStartRunoffSubmission(token, payload)
-		*/
+		// Internal plugin commands
+		// case cmdStartRunoffSubmission:
+		// return p.cmdStartRunoffSubmission(tstore, token, payload)
 	}
 
 	return "", backend.ErrPluginCmdInvalid
